% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotatePeaks.R
\name{annotatePeaks}
\alias{annotatePeaks}
\title{Annotate Regions in the Genome}
\usage{
annotatePeaks(ranges, genome)
}
\arguments{
\item{ranges}{A \code{GRanges} object of peak calls.}

\item{genome}{A character string specifying the genome assembly.}
}
\value{
A \code{GRanges} object with additional metadata columns: \code{entrezTSS}, \code{symbolTSS}, \code{distanceTSS}, and \code{genomicFeature}.
}
\description{
Annotate regions in the genome with nearest TSS and genomic feature.
}
