% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callPeaks.R
\name{callPeaks}
\alias{callPeaks}
\title{Call peaks from DamID-seq data}
\usage{
callPeaks(object, alpha = 0.05, lfc = 0)
}
\arguments{
\item{object}{A RangedSummarizedExperiment object containing normalised counts.}

\item{alpha}{Threshold adjusted P value for declaring whether a peak is statistically significant.}

\item{lfc}{A non-negative value which specifies a log2 fold change threshold.}
}
\value{
A GRanges object of peak regions.
}
\description{
Test each restriction fragment for differential methylation and combine into peak regions.
}
