% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centerMotif.R
\name{centerMotif}
\alias{centerMotif}
\title{Center regions by motif}
\usage{
centerMotif(object, size = NULL)
}
\arguments{
\item{object}{A \code{GRanges} object of peak regions.}
}
\value{
A \code{GRanges} object of peak regions centered by motif position.
}
\description{
Center peak regions by the position of the known motif.
}
