% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeInput.R
\name{computeInput}
\alias{computeInput}
\title{Compute Methylation for Input}
\usage{
computeInput(object)
}
\arguments{
\item{object}{A RangedSummarizedExperiment object.}
}
\value{
A RangedSummarizedExperiment object.
}
\description{
Compute input restriction fragment methylation values for each Dam library.
}
