% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daimTreat.R
\name{daimTreat}
\alias{daimTreat}
\title{Empirical Bayes Statistics For Differential Expression}
\usage{
daimTreat(fit, lfc = log2(1.2), trend = FALSE, robust = FALSE,
  winsor.tail.p = c(0.05, 0.1))
}
\description{
Computes empirical Bayes moderated-t p-values relative to a minimum meaningful fold-change threshold.
}
\details{
This is an internal function adapted from the \code{treat} function in the \code{limma} package. It is used to test for a log2 fold change value (Dam-fusion/Dam) greater than the specified threshold.
}
