% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/digestGenome.R
\name{digestGenome}
\alias{digestGenome}
\title{Restriction digest of genome}
\usage{
digestGenome(object)
}
\arguments{
\item{object}{A \code{BSgenome}, \code{DNAStringSet}, or \code{FaFile} object.}
}
\value{
A \code{\link[GenomicRanges]{GRanges}} object of DpnII restriction fragments.
}
\description{
Extract DpnII restriction fragments from a genome.
}
