% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterBySize.R
\name{filterBySize}
\alias{filterBySize}
\title{Filter Ranges By Size}
\usage{
filterBySize(object, min = 100, max = 1e+05)
}
\arguments{
\item{object}{A GRanges object.}

\item{min}{Minimum size required.}

\item{max}{Maximum size required.}
}
\description{
Filter Ranges By Size
}
