% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fragmentCounts.R
\name{fragmentCounts}
\alias{fragmentCounts}
\title{Count reads in restriction fragments}
\usage{
fragmentCounts(reads, fragments, mode = c("inner", "flank"))
}
\arguments{
\item{reads}{A character vector of paths to BAM files.}

\item{fragments}{A \code{GRanges} object of restriction fragments.}

\item{mode}{Read counting mode. Can be either "inner" or "flank".}
}
\value{
A \code{RangedSummarizedExperiment} object.
}
\description{
Assign mapped sequencing reads to restriction fragments.
}
