% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalizeBias.R
\name{normalizeBias}
\alias{normalizeBias}
\title{Normalize restriction fragment biases}
\usage{
normalizeBias(object, group)
}
\arguments{
\item{object}{A RangedSummarizedExperiment object.}

\item{group}{A factor specifying Dam and Dam-fusion groups.}
}
\value{
A \code{\link[SummarizedExperiment]{RangedSummarizedExperiment}} object.
}
\description{
Perform normalization to remove restriction fragment biases.
}
