% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotComplexity.R
\name{plotComplexity}
\alias{plotComplexity}
\title{Plot restriction fragment complexity}
\usage{
plotComplexity(object, group = NULL)
}
\arguments{
\item{object}{A RangedSummarizedExperiment object.}

\item{group}{A grouping factor.}
}
\value{
A restriction fragment complexity plot on the current graphics device.
}
\description{
Produce a restriction fragment complexity plot.
}
