% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCorrelation.R
\name{plotCorrelation}
\alias{plotCorrelation}
\title{Plot correlation between two samples}
\usage{
plotCorrelation(object, contrast = NULL)
}
\arguments{
\item{object}{A \code{RangedSummarizedExperiment} object.}

\item{contrast}{The name or index of two samples in the object.}
}
\description{
Produce a correlation plot between two samples.
}
