% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCoverage.R
\name{plotCoverage}
\alias{plotCoverage}
\title{Cumulative distribution of sequencing coverage}
\usage{
plotCoverage(object, group = NULL)
}
\arguments{
\item{object}{A \code{RangedSummarizedExperiment} object.}

\item{group}{A factor assigning samples to groups.}
}
\value{
A cumulative distribution of sequence coverge plot on the current graphics device.
}
\description{
Produce a cumulative distribution of sequence coverge plot.
}
