% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDensity.R
\name{plotDensity}
\alias{plotDensity}
\title{Plot Abundance Densities}
\usage{
plotDensity(object, group = NULL)
}
\arguments{
\item{object}{A \code{matrix} or \code{RangedSummarizedExperiment} object.}

\item{group}{A factor specifying group colours.}
}
\value{
An abundance densitites plot on the current graphics device.
}
\description{
Plot the density of abundance values for multiple samples on the same plot.
}
