% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotEnrichment.R
\name{plotEnrichment}
\alias{plotEnrichment}
\title{Distribution of reads along the genome}
\usage{
plotEnrichment(object, group = NULL)
}
\arguments{
\item{object}{A \code{RangedSummarizedExperiment} object.}

\item{group}{A factor specyfing columns of object to groups.}
}
\value{
An enrichment plot on the current graphics device.
}
\description{
Produce a cumulative sum curve showing the fraction of reads aligned to all restriction fragments in the genome.
}
