% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMA.R
\name{plotMA}
\alias{plotMA}
\title{Brand-Altman plot}
\usage{
plotMA(object, contrast = NULL)
}
\arguments{
\item{object}{A \code{RangedSummarizedExperiment} object.}

\item{contrast}{A character vector with exactly two elements: the name or index of the numerator level for the fold change, and the name or index of the denominator level for the change.
#' @return A Brand-Altman plot on the current graphics device.}
}
\description{
Produce a plot of intensity versus difference.
}
