% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMDS.R
\name{plotMDS}
\alias{plotMDS}
\title{Multidimensional Scaling}
\usage{
plotMDS(object, group)
}
\arguments{
\item{object}{A \code{matrix} or \code{RangedSummarizedExperiment} object.}

\item{group}{A factor specifying group colours.}
}
\value{
A multidimensional scaling plot on the current graphics device.
}
\description{
Plot samples on principal coordinates.
}
