% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPCA.R
\name{plotPCA}
\alias{plotPCA}
\title{Principal Components Analysis}
\usage{
plotPCA(object, group = NULL)
}
\arguments{
\item{object}{A \code{matrix} or \code{RangedSummarizedExperiment} object.}

\item{group}{A factor specifying group colours.}
}
\value{
A principal components plot on the current graphics device.
}
\description{
Performs a principal components analysis on the given data and plots the results.
}
