% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTSS.R
\name{plotTSS}
\alias{plotTSS}
\title{plotTSS}
\usage{
plotTSS(object, mode = c("absolute", "relative"))
}
\arguments{
\item{object}{A \code{GRanges} object of annotated peak regions.}

\item{mode}{Plot "absolute" or "relative" distance.}
}
\description{
Plot the distance between peak regions and their putatively regulated genes.
}
