% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rollAssay.R
\name{rollAssay}
\alias{rollAssay}
\title{Summarize counts from neighbouring ranges}
\usage{
rollAssay(object, size = NULL)
}
\arguments{
\item{object}{A RangedSummarizedExperiment object.}

\item{size}{Size allowed between ranges.}
}
\value{
A matrix of summarized counts.
}
\description{
Compute the average count from neighbouring ranges.
}
