% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchMotif.R
\name{searchMotif}
\alias{searchMotif}
\title{Known motif search}
\usage{
searchMotif(object, ranges, matrix)
}
\arguments{
\item{object}{A \code{BSgenome}, \code{DNAStringSet}, or \code{FaFile} object.}

\item{ranges}{A \code{GRanges} object of peak regions.}

\item{matrix}{A position weight \code{matrix} of the known motif.}
}
\value{
A \code{GRanges} object with additional metadata columns:
\item{motifCenter}{The position of the motif center.}
\item{motifScore}{The score of the motif match.}
}
\description{
Search peak regions for known motifs using a position weight matrix.
}
