% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeAssay.R
\name{writeAssay}
\alias{writeAssay}
\title{Write assay to a bigWig file}
\usage{
writeAssay(object, file = "", sample = NULL)
}
\arguments{
\item{object}{A \code{RangedSummarizedExperiment} object.}

\item{file}{A character string naming a file open for writing.}

\item{sample}{Either a column name or index.}
}
\description{
Export assay values over ranges to a bigWig file format.
}
