% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeBroad.R
\name{writeBroad}
\alias{writeBroad}
\title{Write a GRanges object as a broadPeak file}
\usage{
writeBroad(object, file = "")
}
\arguments{
\item{object}{A GRanges object to be written in broadPeak format.}

\item{file}{A character string naming a file open for writing.}
}
\description{
\code{writePeak} prints a GRanges object in broadPeak format to a file.
}
