% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeRatio.R
\name{writeRatio}
\alias{writeRatio}
\title{Write ratio to a bigWig file}
\usage{
writeRatio(object, file = "", contrast = NULL)
}
\arguments{
\item{object}{A \code{RangedSummarizedExperiment} object.}

\item{file}{A character string naming a file open for writing.}

\item{contrast}{A character vector with exactly two elements: the name or index of the numerator level for the fold change, and the name or index of the denominator level for the change.}
}
\description{
Export ratio values over ranges to a bigWig file format.
}
